put env vars tests into new file to fix timeout
